requirements.txt
langchain-core
langchain-community
streamlit
python-dotenv